Fix build issues and update CI infrastructure to detect them (#5083)
* Refs #21387: Update static edp example to build on windows


* Refs #21387: Apply fix on windows colcon meta


* Refs #21387: Add warnings as errors in main CMakeLists.txt


* Refs #21387: Add no-arguments build step


* Refs #21387: Fix warning without security and log_info


* Refs #21387: Set WError only if not TSan


* Refs #21307. Fix warnings on topic_instances example.


* Refs #21307. Update toupper on UnitsParser.


---------

+++ examples/cpp/static_edp_discovery/PublisherApp.hpp
@@ -78,7 +78,7 @@ private:
 
     TypeSupport type_;
 
-    int16_t matched_;
+    int32_t matched_;
 
     uint16_t samples_;
